Bumping version number and date on spec
Bumping the spec version number and date in anticipation of
the 1.8.0 release.  Still waiting on content from Vass before
snapshotting it into the release.

Also, there seems to be a long-standing formatting issue in one
of the first examples in the spec (caught it while scanning
examples as a result of Tom's change), so I fixed the bug, though
more could be done to make this look nice (spacing/font-size
are a bit weird).



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22081 3a8e244f-b0f2-452b-bcba-4c88e055c3ca