Adding some final changes to CONTRIBUTORS/CHANGES
Got late-arriving confirmation from Michael that Joe Yan should be
listed as a contributor to the current release for his contributions,
so added him back to current CONTRIBUTORS.

Also added a second entry about --llvm-wide-opt to the LLVM-specific
section.

And, it seems I hadn't added my own local changes to point to the new
Task Intents section in the spec, maybe because I was uncertain
whether or not they'd make it in?  Probably I just forgot, though.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22088 3a8e244f-b0f2-452b-bcba-4c88e055c3ca