TRIAGE wrap-up:
Not too much else new to report.  Added the date that the
multilocale/diten/needMultiLocales/coforallon_maxThreads non-deterministic
failure occurred on gasnet-everything to the TRIAGE file.
studies/sudoku/dinan/sudoku-smart may have been fixed, but the next TRIAGEr
should check cray-xt.cray-prgenv-gnu.none.fifo.pthreads--linux64.gnu-std just
to be certain.  chpbld02 intel and cray target compiler builds failed this
weekend due to what looks to be an alteration to the compiler modules.  They now
load the xt-libsci/12.1.2 module and as a consequence when the -lsci flag is
thrown, our compiler has problems. Ryan has been informed and will email me back
as to whether this is a permanent or intended change.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22089 3a8e244f-b0f2-452b-bcba-4c88e055c3ca