Uncommenting some .graph files from GRAPHFILES
The chameneos file was previously commented out strictly for aesthetic
reasons; here, I've taken the approach of moving that into a new shootout
appendix section (which we've talked about wanting for other reasons as
well).

The mandelbrot file used to fail for me, but no longer is.  I believe
it may be the case that it was only failing due to the old graphing
capability, now disabled.  So re-enabled that one.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22094 3a8e244f-b0f2-452b-bcba-4c88e055c3ca