Fix XT GASNet build
It seems that something about GASNet's autoconfigure logic has changed
(or else our platforms have changed in a way that causes the logic to
behave differently), as recent builds of GASNet on an XT have failed
when building an mpi-based conduit because we default to segment
everything and GASNet complains that portals can't be built in segment
everything mode.

More specifically, what I think has changed is that I believe when we were
building an MPI conduit before, GASNet would not build the portals conduit,
and now it seems to try to build the portals conduit because it thinks "I
can."  This is similar to behavior we've recently seen with
the aries and gemini conduits.

The fix is similar to what I believe we've done there.  If we're building
anything other than the mpi conduit, we explicitly disable the portals
conduit.  This seems to work when I build GASNet with (a) no
CHPL_COMM_SUBSTRATE set, (b) CHPL_COMM_SUBSTRATE set to portals, (c)
CHPL_COMM_SUBSTRATE set to mpi.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22095 3a8e244f-b0f2-452b-bcba-4c88e055c3ca