Attempt to speed up miniMD
Through some email conversations and sleuthing, our current
hypothesis for the remaining miniMD timeouts in the testing
system is that we're suffering from a combination of:

(a) nested parallelism causing too many tasks for high core-count
    machines

combined with:

(b) compiler configurations for which we're implementing atomics
    using sync vars rather than native instructions

By turning off dataParIgnoreRunningTasks, I'm dealing with the
first issue and seeing noticably faster running times as a result.
Checking this in to see the impacts on nightly timeouts as well
as performance testing.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22097 3a8e244f-b0f2-452b-bcba-4c88e055c3ca