Add abs function for real(64) and a test that fails without it
[Reviewed by Brad, Sung]

Revision 16736 removed a bunch of 64 bit math functions and replaced them with
extern declarations.  That was good for all of the functions that were named
the same in Chapel and C, but the Chapel abs function is named fabs in C.
This causes abs(real) to promote to the complex version which can cause
wrong answers in some situations.  Added an inline abs(real) function that
passes its argument to the extern fabs function.  Added a test that gets a bad
answer without this change.

We should probably add abs(imag), but I haven't done that here.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22102 3a8e244f-b0f2-452b-bcba-4c88e055c3ca