Removing slow nbody cases
A bunch of the original nbody variations have timed out for about as
long as they've existed.  Reviewing them, I think they do reasonable
things, but are not inherently valuable enough to spend the timeout
time on every time we execute performance testing, so am retiring them
as performance tests.  They will remain as correctness tests, at least
for the time being.  The features they exercise may be worth tracking
as tests on a smaller problem size, but I'm not doing that here.

I also added a README to the directory to try and capture the
differences between versions.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22110 3a8e244f-b0f2-452b-bcba-4c88e055c3ca