Don't throw -j to aprun where it will not be understood.
[Reviewed by sungeun.]

My r21441 commit had the unfortunate side effect of causing us to throw
the aprun -j option even on Cray XT systems, where it is not understood
(because not needed due to lack of hyperthreading).  Undo that.  Now we
only throw -j to aprun when "cnselect -l" lists cpus_per_cu as a compute
node attribute.  Since cnselect is designed to work as a compute node
selector (thus its name) for aprun, this should be a sturdy way to tell
if the aprun we're working with will accept -j.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22113 3a8e244f-b0f2-452b-bcba-4c88e055c3ca