Fix more perl die invocations in chplenv scripts.
[Reviewed by Lydia.]

This changes 3 more die invocations in the util/chplenv scripts to warn
invocations, so that the messages come out and we don't get the internal
perl error about an uninitialized string.

(I think the remaining use of die in check-tasks-threads-compatibility
should stay as it is.  It won't cause death or a message if the script
is run incorrectly by another script, but that's not the scenario we're
primarily protecting against.  If the script is run directly with a bad
command line, it does produce the desired message and non-zero exit
status.)

While here, I also removed an extraneous semicolon and replaced a couple
of <TAB>s with equivalent strings of spaces.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22117 3a8e244f-b0f2-452b-bcba-4c88e055c3ca