Replace explicit {0,0} initializer with symbol, for locales in wide pointers.
[Reviewed by Sung.]

We have been using an explicit {0,0} initializer for the locale part of
a wide pointer.  This only works when the locale, a chpl_localeID_t, is
a two-element composite object.  To allow the locale model more freedom
to define chpl_localeID_t however it wants, have it also #define the
initializer value, with a standardized name, and change the compiler to
use that.

The specific file changes are as follows.

runtime/include/localeModels/flat/chpl-locale-model.h
runtime/include/localeModels/numa/chpl-locale-model.h
  #define CHPL_LOCALEID_T_INIT as the initializer for a chpl_localeID_t
  object.

compiler/AST/symbol.cpp
  Use CHPL_LOCALEID_T_INIT instead of {0,0} as the initialization value
  for the locale parts of wide pointers.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22118 3a8e244f-b0f2-452b-bcba-4c88e055c3ca