Add a future about using the C complex types and math functions.
We currently use homegrown complex types rather than relying on the complex
types provided in C. This makes it more difficult to use C complex math
routines, so we implement our own.  Ours are not always perfect as this test
demonstrates.  The way to fix this future is to start using the C complex types
and math routines.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22120 3a8e244f-b0f2-452b-bcba-4c88e055c3ca