Add --report-optimized-loop-iterators flag
[Reviewed by Lydia]

This commit implements the --report-optimized-loop-iterator flag that
when thrown prints out information about any loop that is identified
for the single loop/yield optimization.

While in the iterator code, I clarified and corrected the existing
comments regarding the optimization, and added a global comment at the
top regarding the purpose of the functions in iterator.cpp.

Added a couple of new tests that make sure the optimization is working
by throwing --report-optimized-loop-iterator.  I would like to add a
test that makes sure the optimization fires in certain cases in the
modules (e.g., unstrided, bounded ranges), but I want to come up with
a way that doesn't rely on us updating this test every time the module
being tested change.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22122 3a8e244f-b0f2-452b-bcba-4c88e055c3ca