Changing chpbld02 from testing cray-xt to cray-xe
The module environment on chpbld02 has been set to Cray XE for quite
some time (as I understand it), yet we've been testing the cray-xt
configuration.  It's likely that this doesn't make a difference for
the kind of whitebox testing we're doing (single-locale only), but
I'm updating the string to make it more accurate.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22124 3a8e244f-b0f2-452b-bcba-4c88e055c3ca