In the flat locale model, a global locale ID does not need a sublocale.
[Reviewed by Sung.]

The flat locale model does not have sublocales, so it does not need a
sublocale member in its chpl_localeID_t struct.  We used to have to have
one anyway, because the compiler contained an assumption (in its wide
pointer support) that chpl_localeID_t was a 2-element composite data
type.  We removed that assumption in r22118, so now we can clean up the
chpl_localeID_t type for the flat locale model.  Specifically, here we
remove the (unused) ignore_subloc member, and adjust the symbolic
CHPL_LOCALEID_T_INIT initializer value to just have one value (for the
node).

On the side, I also converted the node member from a 32-bit bitfield of
plain int type to a non-bitfield member of type int32_t.  It was only a
bitfield as a result of an earlier effort to make the unused sublocale
field be just 1 bit long.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22125 3a8e244f-b0f2-452b-bcba-4c88e055c3ca