Add hwloc to the third-party directory
[reviewed by Greg]

This commit adds hwloc v 1.7.2 to the third-party directory.

We're still discussing how will we incorporate hwloc into the chplenv
stuff, so for now if you want to use it with qthreads, you can build
it in the third-party directory and point the qthreads build at it as
described in README.localeModels.  The hwloc installation lands in:

    $CHPL_HOME/third-party/hwloc/install/$CHPL_TARGET_PLATFORM-$CHPL_TARGET_COMPILER




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22130 3a8e244f-b0f2-452b-bcba-4c88e055c3ca