Adding support for running nightly with the hello* examples only
[trivial, not reviewed]

Prior to this commit, if running nightly in parallel, it wasn't
possible to use the 'hello*' tests only.  This commit permits
that, though since those tests are all in one directory, note
that it won't actually result in any parallelism.  This is
still useful, however, to just make sure that the general
threading of the scripts from one to the next works without
having to test a larger test suite like examples.

I think the only reason we hadn't done this previously was
laziness.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22132 3a8e244f-b0f2-452b-bcba-4c88e055c3ca