Factor zip blocks. [reviewed by vass]
Factored out construction of the body of the various compiler-generated
functions that implement zippered iteration.  This removes a bit of verboseness
from the code, but is primarily intended as a debugging aid.  

With the change, it is clear that the new body contains only the statements
added in the same routine.  It is also clear where the body is inserted in the
tree, so a developer can set a breakpoint there and examine it to verify that it
is complete and correct.

[tested in test/release/examples/primers with --verify]


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22133 3a8e244f-b0f2-452b-bcba-4c88e055c3ca