Changed LICENSE organization and updated README w.r.t. LLVM/CLANG
This commit reorganizes the Chapel LICENSE file to conform to a
request by Cray legal to have the BSD license appear monolithically in
a file with no other extraneous text around it (like the text that we
have been using to mention the third-party/ and etc/ directories as an
exception).

To that end, this commit moves the BSD license into a file named
LICENSE.chapel and repurposes the LICENSE file to talk about the
licenses under which both Chapel and its third-party packages are
distributed, pointing to the other files for more information.

I took this opportunity to do something I've been thinking about for
awhile and roll the licenses of the various third-party packages up to
the top level LICENSE file so that if someone was particularly
allergic to a license, they would be made aware of it at the top-level
rather than having to dig deeper into the directory structure.  As
part of this, I also characterized what the different third-party
packages do and under what conditions they're used (since most are
optionally enabled).

While here, I noticed that the license description for LLVM was not as
accurate as it could be, and also that the text about downloading it
on demand was stale.  Updated those.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22134 3a8e244f-b0f2-452b-bcba-4c88e055c3ca