Move up initialization of DefaultRectangular and private copies of defaultDist
[reviewed by David]

This commit moves the explicit initialization of the
DefaultRectangular module earlier in ChapelStandard.  Previously, it
was being initialized by the use statement in the LocaleModel module.
It's possible it can move up even further, but this is the earliest we
need it for now.

This commit also combines the initialization of the locale private
copies of defaultDist with the initialization of the rootLocale,
leaving only the replicated rootLocale initialization in the
InitPrivateGlobals module.  It is our plan to try to merge the
rootLocale replication into the same place or possibly with the task
table initialization in order to remove at least one more forall onto
all the locales during progrKilledrtup.

One test .good file changed due to the module initialization
reordering.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22140 3a8e244f-b0f2-452b-bcba-4c88e055c3ca