Added some more instrumentation to 'nightly' and cleanup.
[reviewed by Lydia]

* Added -retaintree option that does not wipe+recreate the
working tree aka writeable chapel dir. With this option,
svn update is not run on that tree, but make is still run.
(No deep reason, just ease of coding.)

Had to indent some existing code for an extra 'if'.

* Added more printouts telling what is going on.
"Making compiler" and similar messages are a bit redundant,
but show what options are being used - as a reminder.

* Removed repeated checks "if ($status == 0)".
Since 'print' does not update $status AFAIK, this should
result in equivalent code.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22144 3a8e244f-b0f2-452b-bcba-4c88e055c3ca