In 'nightly', tweaked the commit message for perf data files.
[trivial]

This is a follow-up to r22143: added "fields" to conform to the requirements
of the pre-commit hook of the repository at hand.

I added "Description: <empty>" - because if the Description field is omitted
or is actually empty, 'svn log' will duplicate the contents of the Subject
field. Which would be less pretty IMO.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22146 3a8e244f-b0f2-452b-bcba-4c88e055c3ca