Wrapping up last week's triage
There was some noise last week, but mostly minor.  The one thing worth
noting is that the 11 valgrind failures we saw yesterday were somewhat
strange in that they seemed to complete execution (based on the output
of the test matching the .good) but still timed out.  This makes me
wonder whether something may be going wrong with our teardown...  I
didn't have a chance to investigate.

Most of the other diffs here are just me cleaning up various
categories to bring REGRESSIONS in sync with what we're seeing on the
nightly runs and sort things by the order they appear in the logs:




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22157 3a8e244f-b0f2-452b-bcba-4c88e055c3ca