Corrected a format mismatch in debug statements using DO_GET_REFCNT. [reviewed by mferguson]
The format %ld expects a long int.  On 32-bit platforms, this is 32 bits while
on 64-bit platforms, it's 64-bits.  (Actually, it's implementation-defined, but
this is what gcc has chosen w.r.t. the platforms we use.)  The return value from
DO_GET_REFCNT was always a 64-bit unsigned, hence the mismatch on 32-bit
platforms.

The fix is to us a format descriptor that is platform-independent.  These are
provided by <inttypes.h>.  Since the underlying type of the atomic used to
maintain the ref count is uint_least64_t, we use the format specifier for
unsigned 64-bit integers, which is PRIu64.

When the source code is compiled as C++, definition of the standard C format
macros must be requested explicity with the macro __STDC_FORMAT_MACROS, and this
must be defined before the first inclusion of <inttypes.h>.  For present
purposes, this is the inclusion that appears in qbuffer.h, though it may need to
be moved up if its use becomes more widespread.

The compiler error only shows up if CHPL_DEVELOPER is set to true and only if
GCC is used.

[test built using GCC on linux32 and linux64 platforms with CHPL_DEVELOPER=true]




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22161 3a8e244f-b0f2-452b-bcba-4c88e055c3ca