Changing the handling of svn errors on perf data in 'nightly'.
[requested and reviewed by Brad]

Instead of reporting the total number of svn errors, send a notification
email upon each svn failure, just like when a compiler or runtime build fails.

This un-does some of r22143, in particular:
* the inclusion of svn error count in the final email message
  with the details on succeeded/failed tests, and
* the addition of showsystem().

Instead of the above, we now simply execute the svn commands using
the existing mysystem().

While in mysystem(), I also:

* Made $status private to mysystem(). Otherwise it leaked to the caller.
  (The other de-facto locals should also be made private - I did not do
   that to reduce the diffs.)

* Added the fifth formal argument, which, when passed, causes the
  command to be printed out before executing.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22165 3a8e244f-b0f2-452b-bcba-4c88e055c3ca