Adding some tests for task intents in test/parallel/taskPar/taskIntents.
These tests verify that:

* variables of certain types are captured upon begin/cobegin/coforall

* variables of the other types are not captured

* variables are not captured when they are mentioned in a ref clause

* errors are generated when captured variables are assigned to

Since there is a lot of repetition in these tests, I used cpp
to generate most of the .chpl and .good files. All the machinery
is in src/. However, I am checking in the .chpl and .good files
directly, so src/ is not needed to run the tests.

There is some memory corruption/reference counting bug that
prevents certain code from working correctly. I worked around
that with src/*.patch files. I am adding a cross reference
to a .future for that bug:
  test/parallel/begin/vass/begin-dom-arr-mem-corrupt.future

The Makefile does not trace the dependencies on *.patch files
- I did not feel it was worth my time to put that in.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22166 3a8e244f-b0f2-452b-bcba-4c88e055c3ca