Clean up task table initialization
[reviewed by Greg]

This commit does 2 things.  First, change the forall loop used for
initializing the task table into a coforall.  This enables the
coforall+on optimization.  Note that this means that we need 2 tasks
per locale during initialization.  For some tasking layers, e.g. fifo,
this also implies that at least 2 threads per locale are required.
Also note that we need to use the new ref() clause since we are
modifying an outer variable in the body of the coforall.  This case
was not flagged as an error prior to this change because task intents
are not supported for foralls or coforalls nested in foralls.

Secondly, only initialize the task table if task reporting is
requested via the command line flag.  It is not needed otherwise.

Test changes:

The dataPar2 .good files print out debugging info about forall loops,
so remove the ones that were for the task table initialization.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22169 3a8e244f-b0f2-452b-bcba-4c88e055c3ca