Add suppressions for --fast testing
[reviewed by sung] 

This is to move towards enabling --fast nightly testing 

Since --fast turns off bounds checks, and we have a bunch of tests that are
explicitly looking to see that the bounds checks occurred correctly we need
some way to deal with this. One potential solution was to add something like a
.fastcompopts but that could quickly get messy as you start to want
.baselinecompopts and so on. Instead this patch simply suppresses all the tests
that are looking to see that bounds checks worked correctly. In this manner
we're testing that bounds checks get turned off with --fast. The .fastcompopts
option would allow you to test that bounds checking still works correctly with
optimizations turned on. Both approaches have there merits, but adding
suppressions was a lot easier though it would make sense to add some tests that
throw --fast and --bounds-checks especially once bounds checks don't incur as
much of a performance hit. 

While adding the suppressions I noticed that -suppress was no longer working
for paratest so I fixed that and while I was in paratest I cleaned up
print_help (added suppress, removed comm, alphabetized, and changed the usage
comments at the top to a pointer to print_help.) 


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22170 3a8e244f-b0f2-452b-bcba-4c88e055c3ca