Move rootLocale replication earlier and remove the InitPrivateGlobals module
[reviewed by David]

This patch moves the rootLocale replication (implemented in
chpl_rootLocaleInitPrivate()) into the iterator initOnLocales() used
for Locale initialization similar to what I did with defaultDist in
r22140.  The difference is that the rootLocale replication must happen
*after* the Locale is initialized (after the yield).

This makes the InitPrivateGlobals module unnecessary (deleted).

FYI, this change along with r22140 and r22169 (clean up task table
init), reduces the number of times we go "on" to all locales during
initialization from 4 to 2 (3 if using task reporting).  Greg is
working on getting it down to 1.

Some details:

- The most significant change to chp_rootLocaleInitPrivate() is that
  we now use the primitive for chpl_comm_get directly rather than rely
  on the bulk transfer optimization, because not all the mechanisms
  for parallel iteration and bulk transfer are initialized at this
  point.

- This change also allows us to (properly) declare the Locale arrays
  "private".  Prior to this we were not doing so for bootstrapping
  reasons.  This also gets rid of a wart in the previous
  implementation where we were leaking the array descriptor for
  Locales on locale!=0.

- The .good files for certain test programs that print out debug
  information regarding data parallel operations were changed to
  remove the output that was generated by the forall that was in
  InitPrivateGlobals.

- Refreshed comments



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22171 3a8e244f-b0f2-452b-bcba-4c88e055c3ca