This adds syncing the performance graphs to chapel's sourceforge project-web
[reviewed by vass] 

This patch syncs the performance graphs from chap03, chap04, and bradc-lnx to
sourceforge nightly. The syncPerfGraphs script is essentially a large wrapper
for an rsync command for the current host that authenticates over ssh using my
account (since SF doesn't like group or shared accounts.) Note that this syncs
the dygraph graphs only and not the old graphs at all. This will automatically
sync any updates to the html, javascript, and remove any old graphs. 

In the nightly script the python script is simply called and on failure a
separate mail will be sent out (much like for svn or build failures.) 


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22177 3a8e244f-b0f2-452b-bcba-4c88e055c3ca