Adds .py extension to call to syncPerfGraphs
[trivial, not reviewed]

The script is called syncPerfGraphs.py but I tried just calling syncPerfGraphs.
This resulted in system returning -1 since the program could not be found to
execute it. We divided 256 (shift right by 8) to get the actual error code, but
this isn't what we want to do if system couldn't find the command to execute
since we would be led to believe out status code was -0.00390625, which is
pretty confusing. 


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22181 3a8e244f-b0f2-452b-bcba-4c88e055c3ca