Improve comments for chpl_comm_broadcast* routines
[not reviewed, trivial]

In reviewing the Chapel communication interface with the MPICH group,
it has become apparent that the previous comments were not sufficient
to implement the interface without digging into an existing
implementation to try and decipher what it was doing.  As I answer
questions for them, I'm trying to improve comments as we go.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22186 3a8e244f-b0f2-452b-bcba-4c88e055c3ca