Add a test for --printModuleInitOrder
This test makes sure the --printModuleInitOrder works.  Unfortunately,
if the module initialization order changes, the output will change,
but hopefully that won't be too often.  It might also alert us to an
unexpected change.

If it gets onerous to maintain this test, I can come up with something
less annoying.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22197 3a8e244f-b0f2-452b-bcba-4c88e055c3ca