Fix code generation of imaginary literals
[reviewed by hilde]

Prior to this commit, we had a fragile bug which can be illustrated by
the following test:

  extern proc sizeof(e): ssize_t;

  writeln(sizeof(1.0i));
  writeln(sizeof(1.2i));
  writeln(sizeof(2.0i));
  writeln(sizeof(3i));

The correct answer should be "8" for each of these writeln()s, but the
actual result was much more surprising (answer at the end), and the
reason for the answer similarly surprising (answer also at the end).

This patch addresses this by ensuring that the C name of an imaginary
literal is always stored as a floating point string.  This is
necessary for imaginary literals but not real literals because of the
fact that floating point literals may look integral once the "i" is
removed from them.

Added a test that contains the cases above (and a few others) and
de-futurized the floatLiteralSizes test I added yesterday.

----

Answers:

A1: The actual answer obtained was 4 8 8 4.  Why?


A2: "3i" was stored as "3" and codegen'd as sizeof(3).  Since C int
    literals tend to be 32-bits on most compilers, this reported "4".
    Making it "3.0" uses a C double, resulting in "8" as intended.
    "1.2i" and "2.0i" did the right thing for the right reasons.
    "1.0i" fell into the same case as "3i" because the internal
    modules contain a reference to "1i" and our compiler hashes
    literals to avoid storing them over and over again.  Since "1.0i"
    and "1i" represent the same value, the two shared the "1i" entry,
    which was the first one the compiler encountered.  Thus, it was
    code generated as simply "1", similar to the "3i" case.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22202 3a8e244f-b0f2-452b-bcba-4c88e055c3ca