Add support for comm diagnositics during ChapelStandard initialization
[Reviewed by Tom]

This commit adds 2 new config params that enable comm diagnostics
during the initialization of ChapelStandard:

  - printInitVerboseComm turns on verbose comm during init
  - printInitCommCounts turns on comm counts and print them at the end
    of init

This was done by adding 2 new modules to ChapelStandard: one at the
beginning to turn diags on and one at the end to turn them off and
print output if requested.  This required some .good changes, and also
renaming of tests name writeln.chpl (known bug, that is just now being
tickled).  FYI, one such test is in from the spec.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22203 3a8e244f-b0f2-452b-bcba-4c88e055c3ca