Make the output of our machine and config cron scripts more readable.
[reviewed by Sung]

This is a tweak on r21406:

* For information messages, e.g. "Starting XXX on HOST", print the date
consistently at the beginning of the line.

* Drop 'module list; printenv; echo' that was added in r21406.
These scripts fire up 'nightly', and 'nightly' prints that anyway.

A couple of odd balls on the above:

* Add an extra space after the date, to separate it out visually.
So all echos consistently start with: echo `date` ""
I am not attached to that, though - just giving it a try.

* I was not completely sure we did not need 'module list; printenv'
in the scripts for chpbld01,02 - so I left that in.

* I moved the initial 'echo' earlier in the nightly.tcmalloc script
- because now that line does not need to wait for modules to be set up.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22209 3a8e244f-b0f2-452b-bcba-4c88e055c3ca