Thread idle spin instead of spin or cond [reviewed by Brad]
When implementing the change that made sync variables decide whether to use
spinning or conditionals to wait for a result, I had also caused threads
to do the same thing (r22137).  As we have discovered upon analysis, this leads
to a potential situation where we can have a thread lost to waiting on a
conditional when the run time no longer thinks any threads are doing so.
In addition, some of the performance regressions that occurred as a result of
r22137 were actually caused by this aspect of the change.  This commit will
revert to having the threads only spin while waiting for work.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22210 3a8e244f-b0f2-452b-bcba-4c88e055c3ca