Fix verification failure due to passing atomics by reference
[reviewed by hilde]

This commit fixes a verification failure that I introduced in r21924
which Vass introduced a workaround for in r21936.  The issue was that
function calls introduced after function resolution (like
autoDestroys) whose arguments were passed by ref (like autoDestroys of
atomics) were not getting ref temps inserted for their actuals.  The
reason that this didn't actually cause any problems is that the
functions themselves were empty and inlined, so they went away
before codegen.

This fix properly inserts ref temps for these cases, though it's
somewhat unsatisfactory because it seems possible/likely that over
time as we introduce additional function calls late in compilation,
we'll trip over similar issues.

I also pull out Vass's workaround in this commit.

Editorializing for a moment, my personal preference for the long-term
would be to not insert reference temps until very late in compilation
(like right before codegen) under the argument that we might be code
generating to a language that doesn't require reference temps as C
does (for example, if someone tried to do a Fortran back-end
someday...?).  This would also have the benefit of simplifying the AST
for a greater part of the compilation.  We could then make the verify
pass check that the formal and actuals matched exactly, or that the
formal was the reference type of the actual type.  The other thing
that suggests this to me is that I don't believe we should be
inserting reference temps for inlined functions (they should just
refer to the actual directly if it's sufficiently simple), but this is
trickier to implement given how early reference temps are inserted at
present.

All that said, that was a bigger change than I had time to consider
this week, so I'm just mentioning it here for future consideration
(and I've had a todo item to get rid of ref temps for inlined
functions for awhile, so that may be the change that catalyzes making
this change).




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22213 3a8e244f-b0f2-452b-bcba-4c88e055c3ca